+++ src/main/java/com/example/sendowl/controller/BoardController.java
@@ -3,7 +3,7 @@ package com.example.sendowl.controller;
 
 import com.example.sendowl.dto.BoardRequest;
 import com.example.sendowl.dto.BoardResponse;
-import com.example.sendowl.entity.Board;
+import com.example.sendowl.entity.board.Board;
 import com.example.sendowl.service.BoardService;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
