+++ src/main/java/com/example/sendowl/entity/comment/Comment.java
@@ -1,9 +1,13 @@
-package com.example.sendowl.entity;
+package com.example.sendowl.entity.comment;
 
+import com.example.sendowl.entity.BaseEntity;
+import com.example.sendowl.entity.board.Board;
 import com.example.sendowl.entity.user.User;
 import lombok.*;
 
 import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @NoArgsConstructor
@@ -29,6 +33,9 @@ public class Comment extends BaseEntity {
     private Long parentId; // 대댓글 self 참조 부분은 미정입니다.
 
     private Long depth;
+
+    @OneToMany(mappedBy = "comment", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    private List<CommentLike> commentLikeList = new ArrayList<>();
 
     @Builder
     public Comment(Board board, User user, String content, Long parentId, Long depth) {
