+++ src/main/java/com/example/sendowl/repository/CommentRepository.java
@@ -1,13 +1,11 @@
 package com.example.sendowl.repository;
 
 
-import com.example.sendowl.entity.Board;
-import com.example.sendowl.entity.Comment;
+import com.example.sendowl.entity.board.Board;
+import com.example.sendowl.entity.comment.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import javax.persistence.EntityManager;
-import javax.persistence.Persistence;
 import java.util.List;
 import java.util.Optional;
