+++ src/main/java/com/example/sendowl/service/BoardService.java
@@ -1,7 +1,7 @@
 package com.example.sendowl.service;
 
 import com.example.sendowl.dto.BoardRequest;
-import com.example.sendowl.entity.Board;
+import com.example.sendowl.entity.board.Board;
 import com.example.sendowl.repository.BoardRepository;
 import com.example.sendowl.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
